+++ src/data/types.ts
@@ -5,5 +5,6 @@ export type ScholarRow = Tables['scholars']['Row'];
 export type ScholarID = ScholarRow['id'];
 export type ProposalRow = Tables['proposals']['Row'];
 export type ProposalID = ProposalRow['id'];
+export type SupporterRow = Tables['supporters']['Row'];
 export type CurrencyRow = Tables['currencies']['Row'];
 export type CurrencyID = CurrencyRow['id'];