+++ src/routes/venues/proposal/[id]/+page.ts
@@ -2,9 +2,17 @@
 export async function load({ parent, params }) {
 	const { supabase } = await parent();
 
-	const { data, error } = await supabase.from('proposals').select().eq('id', params.id).single();
-
+	const { data: proposal, error: proposalError } = await supabase
+		.from('proposals')
+		.select()
+		.eq('id', params.id)
+		.single();
+	const { data: supporters, error: supportersError } = await supabase
+		.from('supporters')
+		.select('scholarid (id, name), message')
+		.eq('proposalid', params.id);
 	return {
-		proposal: data && error === null ? data : null
+		proposal: proposal && proposalError === null ? proposal : null,
+		supporters: supporters && supportersError === null ? supporters : null
 	};
 }